+++ essencium-backend/src/main/java/de/frachtwerk/essencium/backend/security/JwtTokenAuthenticationFilter.java
@@ -21,7 +21,7 @@ package de.frachtwerk.essencium.backend.security;
 
 import de.frachtwerk.essencium.backend.service.JwtTokenService;
 import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.JwtException;
+import io.jsonwebtoken.security.SignatureException;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
@@ -38,6 +38,7 @@ import org.springframework.security.authentication.AuthenticationCredentialsNotF
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.web.authentication.AbstractAuthenticationProcessingFilter;
+import org.springframework.security.web.authentication.session.SessionAuthenticationException;
 import org.springframework.security.web.util.matcher.RequestMatcher;
 
 /** Filter to extract a JWT Bearer token from the request's Authorization header and verify it */
@@ -74,12 +75,18 @@ public class JwtTokenAuthenticationFilter extends AbstractAuthenticationProcessi
                     new AuthenticationCredentialsNotFoundException(
                         "missing authorization header parameter"));
 
+    return getAuthentication(token);
+  }
+
+  public Authentication getAuthentication(String token) {
     try {
       final Claims claims = jwtTokenService.verifyToken(token);
       final Authentication auth = new JwtAuthenticationToken(claims.getSubject(), claims);
       return getAuthenticationManager().authenticate(auth);
-    } catch (JwtException e) {
+    } catch (SessionAuthenticationException e) {
       throw new BadCredentialsException(e.getMessage());
+    } catch (SignatureException e) {
+      throw new BadCredentialsException("invalid token");
     }
   }
 
@@ -94,7 +101,7 @@ public class JwtTokenAuthenticationFilter extends AbstractAuthenticationProcessi
     chain.doFilter(request, response);
   }
 
-  private static String extractBearerToken(String param)
+  public static String extractBearerToken(String param)
       throws AuthenticationCredentialsNotFoundException {
     Matcher m = headerParamRegex.matcher(param);
     if (!m.find() || m.groupCount() != 1) {