+++ essencium-backend/src/main/java/de/frachtwerk/essencium/backend/service/JwtTokenService.java
@@ -30,6 +30,8 @@ import de.frachtwerk.essencium.backend.security.SessionTokenKeyLocator;
 import io.jsonwebtoken.*;
 import jakarta.annotation.Nullable;
 import java.io.Serializable;
+import java.time.LocalDateTime;
+import java.time.ZoneOffset;
 import java.util.*;
 import java.util.concurrent.TimeUnit;
 import javax.crypto.SecretKey;
@@ -150,7 +152,14 @@ public class JwtTokenService implements Clock {
       String userAgent,
       @Nullable SessionToken refreshToken) {
     if (sessionTokenType == SessionTokenType.ACCESS && refreshToken != null) {
-      sessionTokenRepository.deleteAll(sessionTokenRepository.findAllByParentToken(refreshToken));
+      // invalidate all ACCESS_TOKENs that belong to this REFRESH_TOKEN
+      sessionTokenRepository.findAllByParentToken(refreshToken).stream()
+          .filter(sessionToken -> sessionToken.getExpiration().after(now()))
+          .forEach(
+              sessionToken -> {
+                sessionToken.setExpiration(now());
+                sessionTokenRepository.save(sessionToken);
+              });
     }
     Date now = now();
     Date expiration = Date.from(now.toInstant().plusSeconds(accessTokenExpiration));
@@ -207,11 +216,25 @@ public class JwtTokenService implements Clock {
   @Transactional
   @Scheduled(fixedRateString = "${app.auth.jwt.cleanup-interval}", timeUnit = TimeUnit.SECONDS)
   public void cleanup() {
-    sessionTokenRepository.deleteAllByExpirationBefore(now());
+    sessionTokenRepository.deleteAllByExpirationBefore(
+        Date.from(
+            LocalDateTime.now()
+                .plusSeconds(jwtConfigProperties.getMaxSessionExpirationTime())
+                .toInstant(ZoneOffset.UTC)));
   }
 
   @Override
   public Date now() {
     return new Date();
+  }
+
+  public boolean isAccessTokenValid(String refresh, String access) {
+    SessionToken refreshToken = getRequestingToken(refresh);
+    SessionToken accessToken = getRequestingToken(access);
+    try {
+      return Objects.equals(refreshToken, accessToken.getParentToken());
+    } catch (NullPointerException e) {
+      return false;
+    }
   }
 }